requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
-e git+https://github.com/Endershade98/data-analyzer.git@d4ea19a7c0a94909170b84f5f03e9370f315d648#egg=data_analyser
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.2
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.38.2
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
protobuf==6.30.2
pyarrow==20.0.0
pydeck==0.9.1
PyMySQL==1.1.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==80.4.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.40
streamlit==1.45.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
